Refactor source & nav funcs as separate modules.
