+++ lib/fmt.js
@@ -1,0 +1,6 @@
+const multiSlashRE = /\/{2,}/g;
+
+module.exports = {
+  multiSlashRE,
+  fmtPath: str => `/${str}/`.replace(multiSlashRE, '/'),
+};
